move to completed folder
